Add:SweetDelete

+++ src/journal/views/NoteView.jsx
@@ -56,7 +56,26 @@ export const NoteView = () => {
   }
   
   const onDelete = ()=>{
-    dispatch(startDeleteNotes())
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+
+      if (result.isConfirmed) {
+        dispatch(startDeleteNotes());
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
+      }
+    });
+   
   }
 
   return (
